Move all non-actions from StatisticsCommand to library
